# backend-ttgo/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Overvak - Sistema de monitoreo remoto para personas con Alzheimer a través de un wearable.",
    "description": "Documentación oficial de la API RESTful para el ecosistema **Overvak**. \n\nEste backend gestiona la comunicación entre los dispositivos wearables y la aplicación web. \n\n**Módulos del sistema:**\n1. **Autenticación:** Gestión de cuidadores y tokens.\n2. **Pacientes:** Vinculación de dispositivos y perfiles.\n3. **Datos IoT:** Ingesta de sensores, tales como el MAX30102, GPS (NEO6MV2) y acelerómetro.\n4. **Geocercas:** Algoritmos de punto en polígono para zonas seguras.\n5. **Alertas:** Notificaciones de emergencia.",
    "version": "1.2.1"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Introduce tu token así: Bearer <tu_token>"
    }
  },
  "paths": {
...
    "/api/cuidadores/register": {
      "post": {
        "tags": [
          "Autenticación y manejo del cuidador"
        ],
        "description": "Registro de un nuevo cuidador. Envía correo de verificación.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "type": "string",
                  "example": "Alberto"
                },
                "apellidoP": {
                  "type": "string",
                  "example": "Vega"
                },
                "apellidoM": {
                  "type": "string",
                  "example": "Monterrubio"
                },
                "email": {
                  "type": "string",
                  "example": "example@gmail.com"
                },
                "telefono": {
                  "type": "string",
                  "example": "5512345688"
                },
                "password": {
                  "type": "string",
                  "example": "Passw0rd*"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuario creado. Correo enviado."
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "El correo ya está registrado"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/cuidadores/verificar/{token}": {
      "get": {
        "tags": [
          "Autenticación y manejo del cuidador"
        ],
        "description": "Endpoint accedido desde el correo para activar la cuenta.",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Token JWT enviado por email"
          }
        ],
        "responses": {
          "200": {
            "description": "HTML de confirmación exitosa"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "HTML de error (token inválido)"
          }
        }
      }
    },
    "/api/cuidadores/login": {
      "post": {
        "tags": [
          "Autenticación y manejo del cuidador"
        ],
        "description": "Inicio de sesión. Devuelve Token JWT.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "example@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "pass"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login exitoso",
            "schema": {
              "type": "object",
              "properties": {
                "mensaje": {
                  "type": "string",
                  "example": "Inicio de sesión correcto"
                },
                "token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsIn..."
                },
                "cuidador": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "645..."
                    },
                    "nombre": {
                      "type": "string",
                      "example": "Alberto"
                    },
                    "email": {
                      "type": "string",
                      "example": "example@gmail.com"
                    },
                    "verificado": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Credenciales incorrectas"
          },
          "403": {
            "description": "Cuenta no verificada"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/cuidadores/forgot-password": {
      "post": {
        "tags": [
          "Autenticación y manejo del cuidador"
        ],
        "description": "Solicita un correo para restablecer contraseña.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "example@gmail.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Correo enviado (si existe el usuario)"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/cuidadores/reset-password": {
      "post": {
        "tags": [
          "Autenticación y manejo del cuidador"
        ],
        "description": "Establece una nueva contraseña usando el token recibido por correo.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string",
                  "example": "eyJ..."
                },
                "password": {
                  "type": "string",
                  "example": "NewPassw0rd*"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contraseña actualizada"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Token inválido"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/cuidadores/{id}": {
      "put": {
        "tags": [
          "Autenticación y manejo del cuidador"
        ],
        "description": "Actualiza datos del perfil del cuidador.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del cuidador"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "type": "string",
                  "example": "Alberto Nuevo"
                },
                "telefono": {
                  "type": "string",
                  "example": "5599887766"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Autenticación y manejo del cuidador"
        ],
        "description": "Elimina la cuenta del cuidador.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del cuidador"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}